Internals

Base.isempty — Method

isempty(group::BenchmarkGroup)

Return true if group is empty. This will first run clear_empty! on group to recursively remove any empty subgroups.

BenchmarkTools._withprogress — Method

_withprogress(
    name::AbstractString,
    group::BenchmarkGroup;
    kwargs...,
) do progressid, nleaves, ndone
    ...
end

Execute do block with following arguments:

progressid: logging ID to be used for @logmsg.

nleaves: total number of benchmarks counted at the root benchmark group.

ndone: number of completed benchmarks

They are either extracted from kwargs (for sub-groups) or newly created (for root benchmark group).

BenchmarkTools.load — Method

BenchmarkTools.load(filename)

Load serialized benchmarking objects (e.g. results or parameters) from a JSON file.

BenchmarkTools.quasiquote! — Method

quasiquote!(expr::Expr, vars::Vector{Symbol}, vals::Vector{Expr})

Replace every interpolated value in expr with a placeholder variable and store the resulting variable / value pairings in vars and vals.

BenchmarkTools.save — Method

BenchmarkTools.save(filename, args...)

Save serialized benchmarking objects (e.g. results or parameters) to a JSON file.