Internals

BenchmarkTools._withprogress — Method

_withprogress(
    name::AbstractString,
    group::BenchmarkGroup;
    kwargs...,
) do progressid, nleaves, ndone
    ...
end

Execute do block with following arguments:

progressid: logging ID to be used for @logmsg.

nleaves: total number of benchmarks counted at the root benchmark group.

ndone: number of completed benchmarks

They are either extracted from kwargs (for sub-groups) or newly created (for root benchmark group).

BenchmarkTools.quasiquote! — Method

quasiquote!(expr::Expr, vars::Vector{Symbol}, vals::Vector{Expr})

Replace every interpolated value in expr with a placeholder variable and store the resulting variable / value pairings in vars and vals.